// src/types.ts
export interface Dimensions {
  width: number;
  height: number;
}